+++ zarrs_metadata_ext/src/data_type.rs
@@ -2,6 +2,7 @@
 //!
 //! See <https://zarr-specs.readthedocs.io/en/latest/v3/core/index.html#array-metadata-data-type>.
 
+pub mod fixed_length_utf32;
 pub mod numpy_datetime64;
 pub mod numpy_timedelta64;
 pub mod optional;